Splitting SqlServerTestSuite into 3 files: base, sync, async.
By doing this I found a sync test without an async partner; adding it.
